+++ kexec/arch/i386/kexec-nbi.c
@@ -74,7 +74,7 @@ int nbi_probe(const char *buf, off_t len)
 	struct segheader seg;
 	off_t seg_off;
 	/* If we don't have enough data give up */
-	if ((len < sizeof(hdr)) || (len < 512)) {
+	if (((uintmax_t)len < (uintmax_t)sizeof(hdr)) || (len < 512)) {
 		return -1;
 	}
 	memcpy(&hdr, buf, sizeof(hdr));
