+++ kexec/arch/sh/kexec-uImage-sh.c
@@ -14,7 +14,7 @@ int uImage_sh_probe(const char *buf, off_t len)
 {
 	struct image_header header;
 
-	if (len < sizeof(header))
+	if ((uintmax_t)len < (uintmax_t)sizeof(header))
 		return -1;
 
 	memcpy(&header, buf, sizeof(header));
